requirements.txt
astroid==2.3.2
beautifulsoup4==4.8.1
Django==2.2.6
django-appconf==1.0.3
django-compressor==2.3
html5lib==1.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
ply==3.11
pylint==2.4.3
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
slimit==0.8.1
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
webencodings==0.5.1
wrapt==1.11.2
